# fix: only rebuild proto on changes