--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -46,8 +46,8 @@ jobs:
     name: proto check
     runs-on: ubuntu-latest
     steps:
+      - uses: actions/checkout@main
       - name: Update rust toolchain
-        uses: actions/checkout@main
         run: rustup update --no-self-update
       - name: Rebuild protos
         run: BUILD_PROTO=1 cargo check -p miden-node-rpc -p miden-rpc-proto