--- a/crates/proto/src/generated/rustfmt.toml
+++ b/crates/proto/src/generated/rustfmt.toml
@@ -1,1 +1,0 @@
-disable_all_formatting = true