--- a/crates/rpc-proto/build.rs
+++ b/crates/rpc-proto/build.rs
@@ -22,13 +22,13 @@ const DOC_COMMENT: &str =
 /// This is done only if BUILD_PROTO environment variable is set to `1` to avoid running the script
 /// on crates.io where repo-level .proto files are not available.
 fn main() -> io::Result<()> {
-    println!("cargo:rerun-if-changed=proto");
-    println!("cargo:rerun-if-changed=../../proto");
-
     // skip this build script in BUILD_PROTO environment variable is not set to `1`
     if env::var("BUILD_PROTO").unwrap_or("0".to_string()) == "0" {
         return Ok(());
     }
+
+    println!("cargo:rerun-if-changed=proto");
+    println!("cargo:rerun-if-changed=../../proto");
 
     // copy all .proto files into this crate. all these files need to be local to the crate to
     // publish the crate to crates.io