--- a/crates/rpc-proto/build.rs
+++ b/crates/rpc-proto/build.rs
@@ -1,7 +1,5 @@
 use std::{
-    env,
-    fs::{self, File},
-    io::{self, Read, Write},
+    env, fs, io,
     path::{Path, PathBuf},
 };
 
@@ -22,12 +20,12 @@ const DOC_COMMENT: &str =
 /// This is done only if BUILD_PROTO environment variable is set to `1` to avoid running the script
 /// on crates.io where repo-level .proto files are not available.
 fn main() -> io::Result<()> {
+    println!("cargo:rerun-if-changed=../../proto");
+
     // skip this build script in BUILD_PROTO environment variable is not set to `1`
     if env::var("BUILD_PROTO").unwrap_or("0".to_string()) == "0" {
         return Ok(());
     }
-
-    println!("cargo:rerun-if-changed=../../proto");
 
     // Copy all .proto files into this crate. all these files need to be local to the crate to
     // publish the crate to crates.io
@@ -37,29 +35,30 @@ fn main() -> io::Result<()> {
 
     let out_dir = env::current_dir().expect("Error getting cwd");
     let dest_path = Path::new(&out_dir).join("./src/proto_files.rs");
-    let mut file = File::create(dest_path)?;
 
-    writeln!(file, "/// {DOC_COMMENT}")?;
-    writeln!(file, "pub const PROTO_FILES: &[(&str, &str)] = &[")?;
-
-    for entry in std::fs::read_dir(CRATE_PROTO_DIR)? {
+    let mut proto_filenames = Vec::new();
+    for entry in fs::read_dir(CRATE_PROTO_DIR)? {
         let entry = entry?;
         let path = entry.path();
         if path.is_file() {
-            let mut file_content = String::new();
             let file_name =
                 path.file_name().and_then(|f| f.to_str()).expect("Could not get file name");
 
-            File::open(&path)?.read_to_string(&mut file_content)?;
-            writeln!(
-                file,
-                "    (\"{}\", include_str!(\"../{CRATE_PROTO_DIR}/{}\")),",
-                file_name, file_name
-            )?;
+            proto_filenames.push(format!(
+                "\"{file_name}\" = include_str!(\"../{CRATE_PROTO_DIR}/file_name\",\n"
+            ));
         }
     }
+    // Sort so that the vector is consistent since directory walking order is
+    // not guaranteed, otherwise there will be diffs from different runs.
+    proto_filenames.sort();
 
-    writeln!(file, "];")?;
+    let content =
+        std::iter::once(format!("/// {DOC_COMMENT}\npub const PROTO_FILES: &[(&str, &str)] = &["))
+            .chain(proto_filenames.into_iter())
+            .chain(std::iter::once("];\n".to_string()))
+            .collect::<String>();
+    fs::write(dest_path, content)?;
 
     Ok(())
 }