… crates/proto/src/generated/mod.rs