The proto files are being "rebuilt" on every compile at the moment. This is due to the 
```
    println!("cargo:rerun-if-changed=generated");
```
which is a directory that does not exist, causing it to trigger on every build.

This PR fixes the path to the correct `src/generated` directory and reorganizes the path determination to use the package root instead of the current directory (should be more robust).

I am however unsure why the generated directory is included -- maybe to rebuild it if it accidentally got deleted? Though that seems excessive as the dev can see this happened using git.

### Before
```
❯ cargo check
   Compiling miden-node-proto v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/proto)
    Checking miden-node-store v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/store)
    Checking miden-faucet v0.5.0 (/Users/mvonleipzig/Projects/miden/node/bin/faucet)
    Checking miden-node-block-producer v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/block-producer)
    Checking miden-node-rpc v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/rpc)
    Checking miden-node v0.5.0 (/Users/mvonleipzig/Projects/miden/node/bin/node)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 1.97s

❯ cargo check
   Compiling miden-node-proto v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/proto)
    Checking miden-node-store v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/store)
    Checking miden-faucet v0.5.0 (/Users/mvonleipzig/Projects/miden/node/bin/faucet)
    Checking miden-node-block-producer v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/block-producer)
    Checking miden-node-rpc v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/rpc)
    Checking miden-node v0.5.0 (/Users/mvonleipzig/Projects/miden/node/bin/node)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 1.41s
```

### After
```
❯ cargo check
   Compiling miden-node-proto v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/proto)
    Checking miden-node-store v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/store)
    Checking miden-faucet v0.5.0 (/Users/mvonleipzig/Projects/miden/node/bin/faucet)
    Checking miden-node-block-producer v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/block-producer)
    Checking miden-node-rpc v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/rpc)
    Checking miden-node v0.5.0 (/Users/mvonleipzig/Projects/miden/node/bin/node)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 1.93s

❯ cargo check
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.14s

❯ touch proto/account.proto
❯ cargo check
   Compiling miden-node-proto v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/proto)
   Compiling miden-rpc-proto v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/rpc-proto)
    Checking miden-node-store v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/store)
    Checking miden-faucet v0.5.0 (/Users/mvonleipzig/Projects/miden/node/bin/faucet)
    Checking miden-node-block-producer v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/block-producer)
    Checking miden-node-rpc v0.5.0 (/Users/mvonleipzig/Projects/miden/node/crates/rpc)
    Checking miden-node v0.5.0 (/Users/mvonleipzig/Projects/miden/node/bin/node)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 1.25s
```