@Mirko-von-Leipzig 2024-07-17
As a side note - it feels weird putting "fixes" under `Enhancements` in the changelog. Should there also be a `Fixes`?