@Mirko-von-Leipzig 2024-07-19
I really wish github allowed replying to top level comments..

@bobbinth what about a more explicit approach of moving ownership of the proto definitions to the `proto` crate.  We can then

1. Remove the env flag trigger from the `proto/build.rs` since the definitions are now part of the crate filesystem (I think that was the issue?).
2. Add the functionality required by `rpc-proto` to `proto` under some feature flag (sort of how my current impl merged the two). Really this would only require the actual file contents be made available under a `nostd` implementation.
3. Have `rpc-proto` use `proto` as a dependency.

Benefits:

- A single copy of the proto definitions.
- A single "pure" `build.rs` with no bool flags, no copying around.
- The current (temporary) dependency of `rpc-proto` is made explicit.
- Non-breaking change for both crates.

I'll attempt this so long, but today is a slow day for me it seems so maybe you'll be able to prevent me from wasting my time if you spot any cracks.