@Mirko-von-Leipzig 2024-07-19
I've implemented the latest idea I proposed [here](https://github.com/0xPolygonMiden/miden-node/pull/412#issuecomment-2238882055).

I'm waiting on CI to double-check that I don't need some `nostd` related feature-gating.