@Mirko-von-Leipzig 2024-07-19
We can still add a CI check if we want; but should be difficult to get wrong with the build script