@bobbinth 2024-07-22
I think this approach works, but I'm not sure I like it better than what we had before (assuming we get rid of the un-needed rebuilds).

One reason is that I think there is benefit to keeping `.proto` files separate from the Rust code. These files describe the interfaces of the node components regardless of the language these components are implemented in. Granted, we don't have any plans right now for any non-Rust implementations, but if someone wanted to grab these files to build a Go implementation of the node, it would be a bit more natural to get them from the repo root rather than from one of the internal crates.

Another reason is that as I mentioned, ideally `rpc-proto` shouldn't pull in any of the files related to storage/block producer. We don't have this working yet because of the way we wrote the `.proto` files and refactoring them is a pretty low priority, but it feels a bit hacky to couple `proto` and `rpc-proto` crates.

So, unless there are some complications (i.e., conditional build scripts don't work for some reason), I'd still prefer to keep the `.proto` files in the repo root.