@Mirko-von-Leipzig 2024-07-22
I've reverted to the original format and placing. I've just added a CI check which rebuilds and ensures there is no diff.